Move `src/config` into `src/app/config`
- move `src/config/keys/index.js` into `/config.template.js`
- move `src/config/index.js` into `src/app/config.js`
- merge `initiConfig->group->vendorkeys` within `src/app/dev/index.js`
- remove `src/config/keys/index.template.js`
- deprecate notice for `src/config/keys` folder
- deprecate notice for `src/config` folder
