+++ .gitignore
@@ -16,7 +16,6 @@ package-lock.json
 /src/app/__dev__/
 /src/app/core/iframe/test/
 /src/app/dev/
-/src/config/keys/index.js
 /src/libs/sdk/config/dev/
 /src/libs/sdk/g3w-ol3/config/config.js
 /src/index.html
