+++ src/config/README.md
@@ -1,0 +1,1 @@
+# 🚨 `src/config` folder will be removed in the next major release!
