--- src/config/keys/index.template.js
@@ -1,9 +1,0 @@
-/*
-* You need to copy this file in the same folder named index.js and add your personal keys
-* */
-export const GOOGLE_API_KEY = '<INSERT HERE YOUR GOOGLE API KEY>';
-export const BING_API_KEY = '<INSERT HERE YOUR BING API KEY>';
-export default {
-  GOOGLE_API_KEY,
-  BING_API_KEY
-}
